requirements.txt
Werkzeug
Flask
numpy
gevent
h5py
tensorflow
gunicorn
pillow
matplotlib
keras
opencv-python
imutils